@@ -16,8 +16,8 @@
     "lodash.clamp": "^4.0.3",
     "lodash.shuffle": "^4.2.0",
     "lorem-ipsum": "^2.0.3",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0-alpha-1a3f1afbd",
+    "react-dom": "^18.0.0-alpha-1a3f1afbd",
     "react-feather": "^2.0.9",
     "react-use-gesture": "^9.1.3",
     "react-use-measure": "^2.0.4",
@@ -29,8 +29,8 @@
   "devDependencies": {
     "@types/lodash.clamp": "^4.0.6",
     "@types/lodash.shuffle": "^4.2.6",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.2",
+    "@types/react": "^17.0.11",
+    "@types/react-dom": "^17.0.7",
     "@types/styled-components": "^5.1.9",
     "@vitejs/plugin-react-refresh": "^1.3.1",
     "typescript": "^4.2.3",
