@@ -76,8 +76,8 @@
     "@testing-library/jest-dom": "^4.0.0",
     "@testing-library/react": "^11.2.5",
     "@types/jest": "^26.0.23",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.3",
+    "@types/react": "^17.0.11",
+    "@types/react-dom": "^17.0.7",
     "flush-microtasks": "^1.0.1",
     "husky": "1.3.1",
     "jest": "^26.6.3",
@@ -87,8 +87,8 @@
     "postinstall-postinstall": "^2.1.0",
     "prettier": "^2.2.1",
     "pretty-quick": "1.10.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0-alpha-1a3f1afbd",
+    "react-dom": "^18.0.0-alpha-1a3f1afbd",
     "spec.ts": "^1.1.3",
     "ts-jest": "^26.5.4",
     "typescript": "^4.2.3"
